# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/users": {
            "post": {
                "description": "Register new user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "Register user details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_request.RegisterUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.RegisterUserResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/": {
            "get": {
                "description": "List of users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "List of users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ListUserResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Retrive user by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Retrive user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.RetrieveUserResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.UpdateUserResponseWrapper"
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete user by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.DeleteUserResponseWrapper"
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_joqd_authify_internal_adapter_server_dto_request.RegisterUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "Email of the user\nrequired: true",
                    "type": "string",
                    "example": "some@domain.com"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 2,
                    "example": "john"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "is_staff": {
                    "type": "boolean",
                    "example": false
                },
                "is_superuser": {
                    "type": "boolean",
                    "example": false
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 2,
                    "example": "doe"
                },
                "password": {
                    "description": "Password of the user\nrequired: true",
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8,
                    "example": "strongpassword"
                },
                "username": {
                    "description": "Username of the user\nrequired: true",
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 3,
                    "example": "admin"
                }
            }
        },
        "github_com_joqd_authify_internal_adapter_server_dto_response.DeleteUserResponseWrapper": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.DeletedUser"
                }
            }
        },
        "github_com_joqd_authify_internal_adapter_server_dto_response.DeletedUser": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "github_com_joqd_authify_internal_adapter_server_dto_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "github_com_joqd_authify_internal_adapter_server_dto_response.ListUserResponseWrapper": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.RetrievedUsers"
                }
            }
        },
        "github_com_joqd_authify_internal_adapter_server_dto_response.RegisterUserResponseWrapper": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.RegisteredUser"
                }
            }
        },
        "github_com_joqd_authify_internal_adapter_server_dto_response.RegisteredUser": {
            "type": "object",
            "properties": {
                "date_joined": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_joqd_authify_internal_adapter_server_dto_response.RetrieveUserResponseWrapper": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.RetrievedUser"
                }
            }
        },
        "github_com_joqd_authify_internal_adapter_server_dto_response.RetrievedUser": {
            "type": "object",
            "properties": {
                "date_joined": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_email_verified": {
                    "type": "boolean"
                },
                "is_staff": {
                    "type": "boolean"
                },
                "is_superuser": {
                    "type": "boolean"
                },
                "last_login": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_joqd_authify_internal_adapter_server_dto_response.RetrievedUsers": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.RetrievedUser"
                    }
                }
            }
        },
        "github_com_joqd_authify_internal_adapter_server_dto_response.UpdateUserResponseWrapper": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "$ref": "#/definitions/github_com_joqd_authify_internal_adapter_server_dto_response.UpdatedUser"
                }
            }
        },
        "github_com_joqd_authify_internal_adapter_server_dto_response.UpdatedUser": {
            "type": "object",
            "properties": {
                "date_joined": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_staff": {
                    "type": "boolean"
                },
                "is_superuser": {
                    "type": "boolean"
                },
                "last_login": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}